# services/staffing-service/openapi.yaml
openapi: 3.1.0
info:
  title: CareCareer Staffing Service
  version: 0.1.0
  description: |
    Manages facilities, departments, credential requirements, and workers.
    All endpoints require Bearer token authentication (RS256).
    All operations are tenant-scoped via validated session principal.
servers:
  - url: http://localhost:3200
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /v1/workers:
    post:
      operationId: createWorker
      summary: Create a worker profile
      description: |
        Requires permission: worker.create
        PII fields stored but NEVER logged or included in events.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkerRequest'
      responses:
        '201':
          description: Worker created
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      operationId: listWorkers
      summary: List workers (admin)
      description: Requires permission worker.list
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum:
              [
                APPLICANT,
                SCREENING,
                QUALIFIED,
                CREDENTIALING,
                READY,
                ACTIVE,
                INACTIVE,
                BLOCKED,
                ALUMNI,
              ]
      responses:
        '200':
          description: Worker list
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/workers/{workerId}/credentials:
    post:
      operationId: createCredential
      summary: Add a credential to a worker
      description: Requires permission credentials:create and Idempotency-Key header
      parameters:
        - $ref: '#/components/parameters/WorkerId'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCredentialRequest'
      responses:
        '201':
          description: Credential created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      credentialId:
                        type: string
                        format: uuid
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      operationId: listCredentials
      summary: List credentials for a worker
      description: Requires permission credentials:read
      parameters:
        - $ref: '#/components/parameters/WorkerId'
      responses:
        '200':
          description: Credential list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CredentialSummary'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/workers/{workerId}/credentials/{credentialId}/submit:
    post:
      operationId: submitCredentialForVerification
      summary: Submit credential for compliance review
      description: Transitions UPLOADED -> PENDING_VERIFICATION. Requires Idempotency-Key and expectedVersion.
      parameters:
        - $ref: '#/components/parameters/WorkerId'
        - $ref: '#/components/parameters/CredentialId'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionedMutation'
      responses:
        '200':
          description: Submitted for verification
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/workers/{workerId}/credentials/{credentialId}/verify:
    post:
      operationId: verifyCredential
      summary: Verify a credential (compliance officer action)
      description: Transitions PENDING_VERIFICATION -> VERIFIED. Requires Idempotency-Key and expectedVersion.
      parameters:
        - $ref: '#/components/parameters/WorkerId'
        - $ref: '#/components/parameters/CredentialId'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionedMutation'
      responses:
        '200':
          description: Credential verified
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/workers/{workerId}/credentials/{credentialId}/reject:
    post:
      operationId: rejectCredential
      summary: Reject a credential submission
      description: Transitions PENDING_VERIFICATION -> REJECTED. Requires Idempotency-Key, expectedVersion, and reason.
      parameters:
        - $ref: '#/components/parameters/WorkerId'
        - $ref: '#/components/parameters/CredentialId'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionedReasonRequest'
      responses:
        '200':
          description: Credential rejected
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/workers/{workerId}/credentials/{credentialId}/revoke:
    post:
      operationId: revokeCredential
      summary: Revoke a previously verified credential
      description: Transitions VERIFIED/EXPIRING -> REVOKED. Requires Idempotency-Key, expectedVersion, and reason.
      parameters:
        - $ref: '#/components/parameters/WorkerId'
        - $ref: '#/components/parameters/CredentialId'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionedReasonRequest'
      responses:
        '200':
          description: Credential revoked
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/workers/{workerId}/eligibility-evaluations:
    post:
      operationId: evaluateEligibility
      summary: Evaluate worker eligibility for a facility
      description: Deterministic evaluation at specified checkpoint
      parameters:
        - $ref: '#/components/parameters/WorkerId'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluateEligibilityRequest'
      responses:
        '200':
          description: Eligibility evaluation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EligibilityResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── GP-08: Shifts ───────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256-signed platform access token
  parameters:
...
    WorkerId:
      name: workerId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CorrelationId:
      name: x-correlation-id
      in: header
      required: false
      schema:
        type: string
      description: Request correlation ID for cross-service tracing
    CredentialId:
      name: credentialId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: idempotency-key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
      description: Unique idempotency key for exactly-once semantics
  schemas:
...
    CreateWorkerRequest:
      type: object
      required: [firstName, lastName, email, profession]
      additionalProperties: false
      properties:
        firstName: { type: string, minLength: 1, maxLength: 100 }
        lastName: { type: string, minLength: 1, maxLength: 100 }
        email: { type: string, format: email, maxLength: 254 }
        phone: { type: string, maxLength: 30 }
        profession:
          type: string
          enum: [RN, LPN, CNA, RT, ALLIED]
        specialty: { type: string, maxLength: 100 }
        userId: { type: string, format: uuid }
        homeLatitude: { type: number }
        homeLongitude: { type: number }
        homeCity: { type: string }
        homeState: { type: string }
        homeZip: { type: string }
        externalReferences:
          type: array
          items:
            type: object
            required: [systemName, externalId]
            properties:
              systemName:
                type: string
                enum: [symplr, bullhorn, labor-edge, maestra, auth0]
              externalId: { type: string, minLength: 1, maxLength: 200 }
...
    ErrorEnvelope:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details:
          type: array
          items:
            type: object
            properties:
              path: { type: string }
              message: { type: string }
...
    CreateCredentialRequest:
      type: object
      required: [credentialType]
      additionalProperties: false
      properties:
        credentialType:
          type: string
          minLength: 1
          maxLength: 100
        issuingAuthority:
          type: string
          minLength: 1
          maxLength: 200
        credentialNumber:
          type: string
          minLength: 1
          maxLength: 100
        issuedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
    CredentialSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        credentialType:
          type: string
        status:
          $ref: '#/components/schemas/CredentialStatus'
        issuingAuthority:
          type: string
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        version:
          type: integer
    CredentialStatus:
      type: string
      enum:
        - UPLOADED
        - PENDING_VERIFICATION
        - CORRECTION_REQUIRED
        - VERIFIED
        - EXPIRING
        - EXPIRED
        - REJECTED
        - REVOKED
        - SUPERSEDED
...
    VersionedMutation:
      type: object
      required: [expectedVersion]
      additionalProperties: false
      properties:
        expectedVersion:
          type: integer
          minimum: 1
    VersionedReasonRequest:
      type: object
      required: [expectedVersion, reason]
      additionalProperties: false
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        reason:
          type: string
          minLength: 1
          maxLength: 500
    EvaluateEligibilityRequest:
      type: object
      required: [facilityId, checkpoint]
      additionalProperties: false
      properties:
        facilityId:
          type: string
          format: uuid
        checkpoint:
          type: string
          enum:
            - MARKETPLACE_DISPLAY
            - REQUEST_SUBMISSION
            - ASSIGNMENT_CONFIRMATION
            - CLOCK_IN
    EligibilityResult:
      type: object
      properties:
        data:
          type: object
          properties:
            outcome:
              type: string
              enum: [ELIGIBLE, INELIGIBLE, ELIGIBLE_WITH_EXCEPTION, ERROR]
            checkpoint:
              type: string
            reasons:
              type: array
              items:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  credentialType:
                    type: string
            evaluatedAt:
              type: string
              format: date-time
  responses:
    BadRequest:
      description: Invalid request (validation error or invalid transition)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Authentication failed (missing, invalid, or expired token)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found (or hidden cross-tenant)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Optimistic concurrency conflict (version mismatch)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'